SMART THERMAL HYGROMETER
The objective of the project Smart Thermal Hygrometer using ESP8266 module is to design a product which is very much useful in post COVID-19 era in order to perform activities with no physical contact. Our project model is mainly focused on designing a hygrometer in such a way that the sensed temperature and humidity is displayed and later on analysed and stored into Firebase API which can be accessed through a mobile. However, it not only helps us in reducing corona cases but also helps we performing activities contactless. It also helps to retain data for further analysation